refactor(core): remove ABC-based test functions and abstract classes
Delete legacy ABC-based UQ test functions
and their abstract base classes.
The testing module and corresponding `uqtestfun_abc`
implementations are now obsolete due to migration
to YAML-driven specifications and new core structures.

Closes: #534
Ref: #533
