+++ src/uqtestfuns/__init__.py
@@ -6,12 +6,6 @@ import sys
 
 from .core import Marginal
 from .core import ProbInput
-from .core import (
-    UQTestFunBareABC,
-    UQTestFunABC,
-    UQTestFunFixDimABC,
-    UQTestFunVarDimABC,
-)
 from .core import UQTestFun
 from .core import FunParams
 from .core import RealVariable
@@ -23,10 +17,8 @@ from .test_functions import *  # noqa
 from .meta import UQMetaFunSpec
 from .meta import UQMetaTestFun
 
-from .helpers import list_functions
-
 from . import api
-from .api import create, list_parameters
+from .api import create, list_functions, list_parameters
 
 if sys.version_info >= (3, 8):
     from importlib import metadata
@@ -39,18 +31,14 @@ __all__ = [
     "Marginal",
     "ProbInput",
     "FunParams",
-    "UQTestFunBareABC",
-    "UQTestFunABC",
-    "UQTestFunFixDimABC",
-    "UQTestFunVarDimABC",
     "UQTestFun",
     "RealVariable",
     "test_functions",
     "UQMetaFunSpec",
     "UQMetaTestFun",
-    "list_functions",
     "api",
     "create",
+    "list_functions",
     "list_parameters",
 ]
