+++ src/uqtestfuns/core/__init__.py
@@ -6,12 +6,6 @@ from .design_vars.real_variable import RealVariable
 from .parameters import FunParams, Parameters
 from .prob_input.marginal import Marginal
 from .prob_input.probabilistic_input import ProbInput
-from .uqtestfun_abc import (
-    UQTestFunBareABC,
-    UQTestFunABC,
-    UQTestFunFixDimABC,
-    UQTestFunVarDimABC,
-)
 from .uqtestfun import UQTestFun
 
 __all__ = [
@@ -20,9 +14,5 @@ __all__ = [
     "FunParams",
     "Parameters",
     "RealVariable",
-    "UQTestFunBareABC",
-    "UQTestFunABC",
-    "UQTestFunFixDimABC",
-    "UQTestFunVarDimABC",
     "UQTestFun",
 ]
